Move input parsing into configuration.js
