+++ src/store/modules/configuration.js
@@ -1,3 +1,16 @@
+const parseEventInput = (e) => {
+  if (e.target) {
+    return e.target.type === 'number'
+      ? parseInt(e.target.value, 10)
+      : e.target.value;
+  }
+  if (parseInt(e, 10)) {
+    return parseInt(e, 10);
+  }
+
+  return 0;
+};
+
 export default {
   state: {
     password: '',
@@ -8,23 +21,28 @@ export default {
     minNumeric: 2,
   },
   mutations: {
-    setMinLength(state, length) {
+    setMinLength(state, event) {
+      const length = parseEventInput(event);
       if (typeof length !== 'number' || length < 0) return;
       state.minLength = length;
     },
-    setMinSpecialChar(state, count) {
+    setMinSpecialChar(state, event) {
+      const count = parseEventInput(event);
       if (typeof count !== 'number' || count < 0) return;
       state.minSpecialChar = count;
     },
-    setMinUpperChar(state, count) {
+    setMinUpperChar(state, event) {
+      const count = parseEventInput(event);
       if (typeof count !== 'number' || count < 0) return;
       state.minUpperChar = count;
     },
-    setMinLowerChar(state, count) {
+    setMinLowerChar(state, event) {
+      const count = parseEventInput(event);
       if (typeof count !== 'number' || count < 0) return;
       state.minLowerChar = count;
     },
-    setMinNumeric(state, count) {
+    setMinNumeric(state, event) {
+      const count = parseEventInput(event);
       if (typeof count !== 'number' || count < 0) return;
       state.minNumeric = count;
     },
