+++ src/views/Settings.vue
@@ -3,7 +3,7 @@
     <h1>Settings</h1>
     <div class="input-group">
       <label for="length">Minimum length of password</label>
-      <input id="length" type="number" min="0" :value="config.minLength" @input="setMinLength" />
+      <input id="length" type="number" min="0" :value="config.minLength" @input="setMinLength"/>
     </div>
     <div class="input-group">
       <label for="specialChar">Minimum number of special characters</label>
@@ -38,11 +38,11 @@
     <div class="input-group">
       <label for="numerics">Minimum number of numerical characters</label>
       <input
-      id="numerics"
-      type="number"
-      min="0"
-      @input="setMinNumeric"
-      :value="config.minNumeric" />
+        id="numerics"
+        type="number"
+        min="0"
+        @input="setMinNumeric"
+        :value="config.minNumeric"/>
     </div>
 
     <div class="configurations">
@@ -54,7 +54,8 @@
           :minUpperChar="0"
           :minLowerChar="0"
           :minNumeric="0"
-        >Facebook</ConfigButton>
+        >Facebook
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -62,7 +63,8 @@
           :minUpperChar="3"
           :minLowerChar="3"
           :minNumeric="3"
-        >OSU</ConfigButton>
+        >OSU
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -70,7 +72,8 @@
           :minUpperChar="0"
           :minLowerChar="0"
           :minNumeric="0"
-        >Piazza</ConfigButton>
+        >Piazza
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -78,7 +81,8 @@
           :minUpperChar="1"
           :minLowerChar="1"
           :minNumeric="1"
-        >Github</ConfigButton>
+        >Github
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -86,7 +90,8 @@
           :minUpperChar="1"
           :minLowerChar="1"
           :minNumeric="1"
-        >Youtube</ConfigButton>
+        >Youtube
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -94,7 +99,8 @@
           :minUpperChar="1"
           :minLowerChar="1"
           :minNumeric="1"
-        >Stack Overflow</ConfigButton>
+        >Stack Overflow
+        </ConfigButton>
 
         <ConfigButton
           :minLength="6"
@@ -102,7 +108,8 @@
           :minUpperChar="0"
           :minLowerChar="0"
           :minNumeric="0"
-        >Reddit</ConfigButton>
+        >Reddit
+        </ConfigButton>
 
         <ConfigButton
           :minLength="8"
@@ -110,27 +117,16 @@
           :minUpperChar="2"
           :minLowerChar="2"
           :minNumeric="2"
-        >Default Settings</ConfigButton>
+        >Default Settings
+        </ConfigButton>
       </ul>
     </div>
   </div>
 </template>
 
 <script>
+import { mapMutations } from 'vuex';
 import ConfigButton from '@/components/ConfigButton.vue';
-
-const parseEventInput = (e) => {
-  if (e.target) {
-    return e.target.type === 'number'
-      ? parseInt(e.target.value, 10)
-      : e.target.value;
-  }
-  if (parseInt(e, 10)) {
-    return parseInt(e, 10);
-  }
-
-  return 0;
-};
 
 export default {
   components: {
@@ -142,42 +138,34 @@ export default {
     },
   },
   methods: {
-    setMinLength(e) {
-      this.$store.commit('setMinLength', parseEventInput(e));
-    },
-    setMinSpecialChar(e) {
-      this.$store.commit('setMinSpecialChar', parseEventInput(e));
-    },
-    setMinUpperChar(e) {
-      this.$store.commit('setMinUpperChar', parseEventInput(e));
-    },
-    setMinLowerChar(e) {
-      this.$store.commit('setMinLowerChar', parseEventInput(e));
-    },
-    setMinNumeric(e) {
-      this.$store.commit('setMinNumeric', parseEventInput(e));
-    },
+    ...mapMutations([
+      'setMinLength',
+      'setMinSpecialChar',
+      'setMinUpperChar',
+      'setMinLowerChar',
+      'setMinNumeric',
+    ]),
   },
 };
 </script>
 
 <style lang="scss">
-.settings {
-  display: flex;
-  flex-direction: column;
-
-  .input-group {
+  .settings {
     display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-content: center;
-    flex-wrap: nowrap;
-    margin: 0em 0.5rem 1.5rem 0.5rem;
-    font-size: 1.25em;
+    flex-direction: column;
 
-    input {
-      font-size: inherit;
+    .input-group {
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      align-content: center;
+      flex-wrap: nowrap;
+      margin: 0 0.5rem 1.5rem 0.5rem;
+      font-size: 1.25em;
+
+      input {
+        font-size: inherit;
+      }
     }
   }
-}
 </style>
